feat: add app config connection string and access key to parser, restructure options struct and handling
