+++ environment.go
@@ -1,0 +1,42 @@
+package azcfg
+
+const (
+	// azcfgTenantID is the environment variable for the tenant ID.
+	azcfgTenantID = "AZCFG_TENANT_ID"
+	// azcfgClientID is the environment variable for the client ID.
+	azcfgClientID = "AZCFG_CLIENT_ID"
+	// azcfgClientSecret is the environment variable for the client
+	// secret.
+	azcfgClientSecret = "AZCFG_CLIENT_SECRET"
+	// azcfgClientCertificate is the environment variable for the client
+	// certificate.
+	azcfgClientCertificate = "AZCFG_CLIENT_CERTIFICATE"
+	// azcfgClientCertificatePath is the environment variable for the path
+	// to the client certificate.
+	azcfgClientCertificatePath = "AZCFG_CLIENT_CERTIFICATE_PATH"
+	// azcfgAzureCLICredential is the environment variable for the Azure CLI
+	// credential.
+	azcfgAzureCLICredential = "AZCFG_AZURE_CLI_CREDENTIAL"
+	// azcfgKeyVaultName is the environment variable for the Key Vault name.
+	azcfgKeyVaultName = "AZCFG_KEYVAULT_NAME"
+	// azcfgAppConfigurationName is the environment variable for the App
+	// Configuration name.
+	azcfgAppConfigurationName = "AZCFG_APPCONFIGURATION_NAME"
+	// azcfgAppConfigurationAccessKeyID is the environment variable for the App
+	// Configuration access key ID.
+	azcfgAppConfigurationAccessKeyID = "AZCFG_APPCONFIGURATION_ACCESS_KEY_ID"
+	// azcfgAppConfigurationAccessKeySecret is the environment variable for the App
+	// Configuration access key secret.
+	azcfgAppConfigurationAccessKeySecret = "AZCFG_APPCONFIGURATION_ACCESS_KEY_SECRET"
+	// azcfgAppConfigurationConnectionString is the environment variable for the App
+	// Configuration connection string.
+	azcfgAppConfigurationConnectionString = "AZCFG_APPCONFIGURATION_CONNECTION_STRING"
+	// azcfgAppConfigurationLabel is the environment variable for the App
+	// Configuration label.
+	azcfgAppConfigurationLabel = "AZCFG_APPCONFIGURATION_LABEL"
+	// azcfgAppConfigurationLabels is the environment variable for the App
+	// Configuration labels.
+	azcfgAppConfigurationLabels = "AZCFG_APPCONFIGURATION_LABELS"
+	// azcfgCloud is the environment variable for the (azure) cloud.
+	azcfgCloud = "AZCFG_CLOUD"
+)
