+++ parser_test.go
@@ -48,7 +48,7 @@ func TestNewParser(t *testing.T) {
 				settingClient: &setting.Client{},
 				cred:          &identity.ClientCredential{},
 				timeout:       time.Second * 10,
-				concurrency:   10,
+				concurrency:   defaultConcurrency,
 			},
 		},
 		{
@@ -68,7 +68,7 @@ func TestNewParser(t *testing.T) {
 				secretClient: &secret.Client{},
 				cred:         &identity.ClientCredential{},
 				timeout:      time.Second * 10,
-				concurrency:  10,
+				concurrency:  defaultConcurrency,
 			},
 		},
 		{
@@ -88,7 +88,7 @@ func TestNewParser(t *testing.T) {
 				settingClient: &setting.Client{},
 				cred:          &identity.ClientCredential{},
 				timeout:       time.Second * 10,
-				concurrency:   10,
+				concurrency:   defaultConcurrency,
 			},
 		},
 		{
@@ -98,7 +98,7 @@ func TestNewParser(t *testing.T) {
 				envs    map[string]string
 			}{
 				options: []Option{
-					WithConcurrency(20),
+					WithConcurrency(30),
 					WithTimeout(time.Second * 10),
 					WithClientSecretCredential("1111", "2222", "3333"),
 					WithKeyVault("vault1"),
@@ -110,7 +110,7 @@ func TestNewParser(t *testing.T) {
 				settingClient: &setting.Client{},
 				cred:          &identity.ClientCredential{},
 				timeout:       time.Second * 10,
-				concurrency:   20,
+				concurrency:   30,
 			},
 		},
 		{
@@ -145,7 +145,7 @@ func TestNewParser(t *testing.T) {
 				settingClient: nil,
 				cred:          nil,
 				timeout:       time.Second * 10,
-				concurrency:   10,
+				concurrency:   defaultConcurrency,
 			},
 		},
 		{
@@ -163,7 +163,7 @@ func TestNewParser(t *testing.T) {
 				settingClient: stub.SettingClient{},
 				cred:          nil,
 				timeout:       time.Second * 10,
-				concurrency:   10,
+				concurrency:   defaultConcurrency,
 			},
 		},
 		{
@@ -182,7 +182,7 @@ func TestNewParser(t *testing.T) {
 				settingClient: stub.SettingClient{},
 				cred:          nil,
 				timeout:       time.Second * 10,
-				concurrency:   10,
+				concurrency:   defaultConcurrency,
 			},
 		},
 	}
@@ -231,7 +231,8 @@ func TestSetupCredential(t *testing.T) {
 	var tests = []struct {
 		name  string
 		input struct {
-			options Options
+			cloud   cloud.Cloud
+			options Entra
 			envs    map[string]string
 		}
 		want    auth.Credential
@@ -240,7 +241,8 @@ func TestSetupCredential(t *testing.T) {
 		{
 			name: "credential settings from environment (client secret credential)",
 			input: struct {
-				options Options
+				cloud   cloud.Cloud
+				options Entra
 				envs    map[string]string
 			}{
 				envs: map[string]string{
@@ -254,7 +256,8 @@ func TestSetupCredential(t *testing.T) {
 		{
 			name: "credential settings from environment (client certificate credential from base64)",
 			input: struct {
-				options Options
+				cloud   cloud.Cloud
+				options Entra
 				envs    map[string]string
 			}{
 				envs: map[string]string{
@@ -268,7 +271,8 @@ func TestSetupCredential(t *testing.T) {
 		{
 			name: "credential settings from environment (client certificate credential from file)",
 			input: struct {
-				options Options
+				cloud   cloud.Cloud
+				options Entra
 				envs    map[string]string
 			}{
 				envs: map[string]string{
@@ -282,34 +286,20 @@ func TestSetupCredential(t *testing.T) {
 		{
 			name: "credential settings from environment (managed identity)",
 			input: struct {
-				options Options
+				cloud   cloud.Cloud
+				options Entra
 				envs    map[string]string
 			}{},
 			want: &identity.ManagedIdentityCredential{},
 		},
 		{
-			name: "credential settings from options (provided credential)",
-			input: struct {
-				options Options
-				envs    map[string]string
-			}{
-				options: Options{
-					Credential: mockCredential{
-						t: "custom",
-					},
-				},
-			},
-			want: mockCredential{
-				t: "custom",
-			},
-		},
-		{
 			name: "credential settings from options (client secret credential)",
 			input: struct {
-				options Options
+				cloud   cloud.Cloud
+				options Entra
 				envs    map[string]string
 			}{
-				options: Options{
+				options: Entra{
 					TenantID:     "1111",
 					ClientID:     "2222",
 					ClientSecret: "3333",
@@ -320,10 +310,11 @@ func TestSetupCredential(t *testing.T) {
 		{
 			name: "credential settings from options (client certificate credential)",
 			input: struct {
-				options Options
+				cloud   cloud.Cloud
+				options Entra
 				envs    map[string]string
 			}{
-				options: Options{
+				options: Entra{
 					TenantID:     "1111",
 					ClientID:     "2222",
 					Certificates: []*x509.Certificate{{}},
@@ -335,17 +326,19 @@ func TestSetupCredential(t *testing.T) {
 		{
 			name: "credential settings from options (managed identity)",
 			input: struct {
-				options Options
+				cloud   cloud.Cloud
+				options Entra
 				envs    map[string]string
 			}{
-				options: Options{ManagedIdentity: true},
+				options: Entra{ManagedIdentity: true},
 			},
 			want: &identity.ManagedIdentityCredential{},
 		},
 		{
 			name: "error setting up client certificate credential",
 			input: struct {
-				options Options
+				cloud   cloud.Cloud
+				options Entra
 				envs    map[string]string
 			}{
 				envs: map[string]string{
@@ -355,18 +348,6 @@ func TestSetupCredential(t *testing.T) {
 				},
 			},
 			wantErr: errors.New("invalid path"),
-		},
-		{
-			name: "error missing client ID",
-			input: struct {
-				options Options
-				envs    map[string]string
-			}{
-				envs: map[string]string{
-					azcfgTenantID: "1111",
-				},
-			},
-			wantErr: ErrMissingClientID,
 		},
 	}
 
@@ -403,7 +384,7 @@ func TestSetupCredential(t *testing.T) {
 				t.Setenv(k, v)
 			}
 
-			got, gotErr := setupCredential(test.input.options)
+			got, gotErr := setupCredential(test.input.cloud, test.input.options)
 
 			if diff := cmp.Diff(test.want, got, cmp.AllowUnexported(mockCredential{}), cmpopts.IgnoreUnexported(identity.ClientCredential{}, identity.ManagedIdentityCredential{})); diff != "" {
 				t.Errorf("setupCredential() = unexpected (-want +got)\n%s\n", diff)
@@ -416,283 +397,7 @@ func TestSetupCredential(t *testing.T) {
 	}
 }
 
-func TestCoalesceString(t *testing.T) {
-	var tests = []struct {
-		name  string
-		input struct {
-			x, y string
-		}
-		want string
-	}{
-		{
-			name: "x is not empty",
-			input: struct {
-				x, y string
-			}{
-				x: "x",
-				y: "y",
-			},
-			want: "x",
-		},
-		{
-			name: "x is empty",
-			input: struct {
-				x, y string
-			}{
-				x: "",
-				y: "y",
-			},
-			want: "y",
-		},
-		{
-			name: "x and y are empty",
-			input: struct {
-				x, y string
-			}{
-				x: "",
-				y: "",
-			},
-			want: "",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			got := coalesceString(test.input.x, test.input.y)
-
-			if test.want != got {
-				t.Errorf("coalesceString() = unexpected result, want: %s, got: %s\n", test.want, got)
-			}
-		})
-	}
-}
-
-func TestCoalesceBool(t *testing.T) {
-	var tests = []struct {
-		name  string
-		input struct {
-			x, y bool
-		}
-		want bool
-	}{
-		{
-			name: "x is true",
-			input: struct {
-				x, y bool
-			}{
-				x: true,
-				y: false,
-			},
-			want: true,
-		},
-		{
-			name: "x is false",
-			input: struct {
-				x, y bool
-			}{
-				x: false,
-				y: true,
-			},
-			want: true,
-		},
-		{
-			name: "x and y are false",
-			input: struct {
-				x, y bool
-			}{
-				x: false,
-				y: false,
-			},
-			want: false,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			got := coalesceBool(test.input.x, test.input.y)
-
-			if test.want != got {
-				t.Errorf("coalesceBool() = unexpected result, want: %t, got: %t\n", test.want, got)
-			}
-		})
-	}
-}
-
-func TestCoalesceMap(t *testing.T) {
-	var tests = []struct {
-		name  string
-		input struct {
-			x, y map[string]string
-		}
-		want map[string]string
-	}{
-		{
-			name: "x is not empty",
-			input: struct {
-				x, y map[string]string
-			}{
-				x: map[string]string{
-					"setting1": "prod",
-				},
-			},
-			want: map[string]string{
-				"setting1": "prod",
-			},
-		},
-		{
-			name: "x is empty",
-			input: struct {
-				x, y map[string]string
-			}{
-				x: nil,
-				y: map[string]string{
-					"setting1": "prod",
-				},
-			},
-			want: map[string]string{
-				"setting1": "prod",
-			},
-		},
-		{
-			name: "x and y are empty",
-			input: struct {
-				x, y map[string]string
-			}{
-				x: nil,
-				y: nil,
-			},
-			want: nil,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			got := coalesceMap(test.input.x, test.input.y)
-
-			if diff := cmp.Diff(test.want, got); diff != "" {
-				t.Errorf("coalesceMap() = unexpected result (-want +got)\n%s\n", diff)
-			}
-		})
-	}
-}
-
-func TestParseLabels(t *testing.T) {
-	var tests = []struct {
-		name  string
-		input string
-		want  map[string]string
-	}{
-		{
-			name: "empty string",
-			want: nil,
-		},
-		{
-			name:  "with labels",
-			input: "setting1=prod,setting2=test",
-			want: map[string]string{
-				"setting1": "prod",
-				"setting2": "test",
-			},
-		},
-		{
-			name:  "with labels (spaces in string)",
-			input: "setting1 = prod, setting2 = test",
-			want: map[string]string{
-				"setting1": "prod",
-				"setting2": "test",
-			},
-		},
-		{
-			name:  "with malformed label",
-			input: "setting1",
-			want:  nil,
-		},
-		{
-			name:  "with malformed second label",
-			input: "setting1=prod,setting2",
-			want: map[string]string{
-				"setting1": "prod",
-			},
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			got := parseLabels(test.input)
-
-			if diff := cmp.Diff(test.want, got); diff != "" {
-				t.Errorf("parseLabels() = unexpected result (-want +got)\n%s\n", diff)
-			}
-		})
-	}
-}
-
-func TestParseCloud(t *testing.T) {
-	var tests = []struct {
-		name  string
-		input string
-		want  cloud.Cloud
-	}{
-		{
-			name: "empty string",
-			want: cloud.AzurePublic,
-		},
-		{
-			name:  "Azure",
-			input: "Azure",
-			want:  cloud.AzurePublic,
-		},
-		{
-			name:  "AzurePublic",
-			input: "AzurePublic",
-			want:  cloud.AzurePublic,
-		},
-		{
-			name:  "Public",
-			input: "Public",
-			want:  cloud.AzurePublic,
-		},
-		{
-			name:  "AzureGovernment",
-			input: "AzureGovernment",
-			want:  cloud.AzureGovernment,
-		},
-		{
-			name:  "Government",
-			input: "Government",
-			want:  cloud.AzureGovernment,
-		},
-		{
-			name:  "AzureChina",
-			input: "AzureChina",
-			want:  cloud.AzureChina,
-		},
-		{
-			name:  "China",
-			input: "China",
-			want:  cloud.AzureChina,
-		},
-		{
-			name:  "non-existent cloud",
-			input: "NonExistent",
-			want:  cloud.AzurePublic,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			got := parseCloud(test.input)
-
-			if test.want != got {
-				t.Errorf("parseCloud() = unexpected result, want: %s, got: %s\n", test.want, got)
-			}
-		})
-	}
-}
-
-type mockCredential struct {
-	t string
-}
+type mockCredential struct{}
 
 func (c mockCredential) Token(ctx context.Context, options ...auth.TokenOption) (auth.Token, error) {
 	return auth.Token{}, nil
